refactored native DeepCausality backend in web-ng/native
