+++ elixir/web-ng/native/god_view_nif/Cargo.toml
@@ -12,9 +12,9 @@ arrow-array = "58"
 arrow-ipc =  "58"
 arrow-schema =  "58"
 deep_causality = "0.13"
-deep_causality_sparse = "0.1.6"
-deep_causality_tensor = "0.4.1"
-deep_causality_topology = "0.5.0"
+deep_causality_sparse = "0.1"
+deep_causality_tensor = "0.4"
+deep_causality_topology = "0.5"
 roaring = "0.11"
 rustler = "0.37"
 serde_json = "1.0"
