+++ elixir/web-ng/native/god_view_nif/src/errors/mod.rs
@@ -1,0 +1,1 @@
+// Custom error types or Result mappings go here
