+++ elixir/web-ng/native/god_view_nif/src/traits/mod.rs
@@ -1,0 +1,1 @@
+// Traits goes here if abstractions are required later.
