+++ elixir/web-ng/native/god_view_nif/src/types/causality.rs
@@ -1,0 +1,10 @@
+use rustler::NifMap;
+
+#[derive(Debug, Clone, NifMap)]
+pub struct CausalStateReasonRow {
+    pub state: u8,
+    pub reason: String,
+    pub root_index: i64,
+    pub parent_index: i64,
+    pub hop_distance: i64,
+}
