+++ elixir/web-ng/native/god_view_nif/src/types/telemetry.rs
@@ -1,0 +1,4 @@
+#[derive(Debug, Clone, Default)]
+pub struct InterfaceTelemetryRecord {
+    pub speed_bps: u64,
+}
